requirements.txt
schedule
python-dotenv
smtplib